Split ttsl::hash out of reflection.hpp into tt_stl/hash.hpp
Extract the ttsl::hash facility (hash_object/hash_objects/hash_combine,
is_std_hashable, the std::hash<SmallVector> specialization, and the
is_specialization/is_span helpers) out of the reflection.hpp monolith into a
new self-contained tt_stl/hash.hpp, so code can hash via <tt_stl/hash.hpp>
without pulling in reflection (nlohmann/json, enchantum, tt-logger).

hash.hpp carries its own member-detection traits in ttsl::hash::detail and
does not include reflection.hpp. The dependency runs one way only:
reflection.hpp includes hash.hpp (Attribute::to_hash_impl_ needs
hash::detail::hash_object, and consumers reach ttsl::hash::* transitively
through <tt_stl/reflection.hpp>). Verbatim move, no behavior change.
