+++ tt_stl/CMakeLists.txt
@@ -18,6 +18,7 @@ target_sources(
             tt_stl/cleanup.hpp
             tt_stl/concepts.hpp
             tt_stl/fmt.hpp
+            tt_stl/hash.hpp
             tt_stl/indestructible.hpp
             tt_stl/optional_reference.hpp
             tt_stl/overloaded.hpp
